count neighbours also counts double items
it can be that a machine in a different subnet or zone has the same
address as another. There is no distinction between these in the
meshnet config (it is by address, not pubkey) because it does not
matter in that context, but it does matter when deciding when to try to
bootstap the distributed network. This commmit adds a function that
counts the neighbours (not unique) and excludes the public key that
belongs to the caller.