implement 'raptiformica ssh'
Get the first available shell or print an ssh command to connect over ssh

```
$ raptiformica ssh
Warning: Permanently added '172.17.0.6' (ECDSA) to the list of known hosts.
root@379683bbc562:~# logout
Connection to 172.17.0.6 closed.
$ raptiformica ssh --info-only
ssh root@172.17.0.3 -p 22 -oStrictHostKeyChecking=no -oUserKnownHostsFile=/dev/null -oPasswordAuthentication=no
```

Currently this is semi-random because python3 lists do not have order. That's fine. In the future I'd like to
implement filters that get you a shell on the 'fastest' machine in the cluster, or the first machine tagged with
'disposable' or something.