notify network reload
when a new machine is added to the distributed kv store config

- uses a consul watcher that will reload cjdroute and consul if necessary
- tries to send the event on one of its peers so the command line
  executor does not have to be part of the network

this is needed for when a machine which is not in the network but has a
config that can talk to machine b which is in the network tries to join
machine c which can not reach b directly but can be reached by b. b will
then create a reverse tunnel to c so that c then can connect to b
transparently as if it could reach it directly.