TreeCluster test scenario
This test case simulates a situation where instance A can reach B using an address that
only has meaning to A. The same goes for A to C. This means that the address that A uses
to connect to B can not be used by C to connect to B. The address used by A to connect to C
can not be used by B to connect to C. A is unreachable for both B and C.

The only path is then:

.===.   .===.   .===.
| A |->-| B |   | C |
'=|='   '==='   '-|-'
  '=======>======='

This scenario tests whether starting from A, slaving B will create a reverse
route using CJDNS from B to A and that slaving C from A will create another
reverse route. C should then be able to connect to B using the two tunnels
and vice versa. This simulates a NAT-like scenario where a firewalled host
can connect two public hosts with each other through the host behind the
firewall. The two public hosts (B and C) do not connect directly.