detect kv content before accepting proxy
only forward the consul HTTP-API port if the remote machine has the
raptiformica_api_version in the kv store, this means there must have
been a quorum in the past and data can be queried through the API.
currently only `consul members` is tested for, which will also return
zero if the KV store can not be accessed. In that case we're better of
trying another machine in the known neighbours list or compose a new
config locally.