do not ping host from same host
When re-assimilating a machine into the network the public key will have
to be updated on a neighbour that can ping the machine. Currently it is
possible that the machine that needs to be updated is regarded as the
first machine that can ping the machine for which a neighbour should be
updated. Instead only all other machines should be tested for being able
to ping the re-assimilated machine.