skip bootstrap new peer if can not ssh
a meshnet reload event will be sent to all connected agents, if this
step fails before a cluster is established the process will fail anyway
later. If it fails after a cluster has already been established the
reload event might still fix up the routes in time.

this situation can happen if there are two machines in the cluster with
guests with the same IPs for example. If machine 1 is spawning a docker
with IP 172.17.0.2 and machine 2 already hosts a docker with that IP and
they're both connected, a failing SSH connection from machine 2 to the
local Docker instance could previously halt the process.