do not block until consul port is free
It is possible that the raptiformica agent already restarted consul
after killing it and before checking if the port is free. In that case,
everything is fine and we shouldn't wait for the consul port to become
free (because it never will unless something kills it again). If the
port is not free and consul is started but can't use the port yet,
that's fine too because the agent will start it later anyway if it
failed. The downside of this is that failure will be less explicit. If
the port really isn't freed for some reason that will not become
immediately apparent anymore.