retry getting shared secret
re-generating the shared secret will likely result in a netsplit. before
we do so let's first retry getting any existing key a couple more times.
downside of this is that it will block for 10s in the retry loop for
each service in case there actually is no secret yet.