add timeout to blocking stdout pipe read
sometimes during high load a shell command will die for various reasons,
in that case the pipe that is being blocking read from by python will
never return anything and the program will block indefinitely. the
threading.Timer runs a separate thread that will kill the process if it
runs out the configured timeout. if required, a @retry decorator can be
added to functions that use exec_command with expected=(TimeoutError,)
so the command will be retried in case it is killed.