add clean state command
Remove all state so it is as if the local machine
will join the cluster for the first time the next
time it is slaved.

Note that this should in theory never be necessary
but it can be helpful for debugging to rule out any
leaking state issues. For example it could be that
one of the instances in the cluster tries to keep
connecting to some stale node that for some reason
isn't being removed from the settings causing the
stability of the mesh to degrade.