integration test concurrently booting clusters
previously the integration tests only tested all cases by booting the
instances one by one. but in real life it is often the case that
multiple nodes come online at the same time. this commit adds new test
cases for all existing integration test cases but where each node in the
test cluster is booted at the same time and where two nodes are booted
at the same time (and the third one whenever the worker finishes with
the first one of the first two).