the default provisioner requires no source
prevents the program from having to clone the raptiformica source into /usr/etc/raptiformica_default_provisioner additionally to /usr/etc/raptiformica. currently this was cloned from /usr/share/raptiformica but that isn't guaranteed to exist because not all hosts in the network will have raptiformica globally installed.  afaik there is also no guarantee that it will (already) exist in /usr/etc/raptiformica at the moment of cloning and there is no way to specify a relative path so it could be retrieved from the ~/.raptiformica.d/artifacts/repos dir.