detach vagrant ssh from terminal
this is horrible for many reasons, but for now it will do the job.
vagrant ssh can't run in the background because it will immediately
receive a SIGTTOU. if you restart the process with a SIGCONT you will
receive another SIGTTOU immediately. this will prevent the command
from freezing even if it is detached. possible side effects: if no
terminal is available wherever this script runs a nohup.out might be
created in the future this oneliner should just become a python script
with proper parsing.

https://www.vagrantup.com/docs/cli/ssh.html#background-execution