disable guest additions in default vagrant
causes problems when booting instances at the same time on the same
host. don't need any of the guest additions features anyway, and if in
the future it is needed then it should just be baked into the boxfile.

still restarts the box on boot to unload any previously loaded vbox
guest additions kernel module