disable semi-concurrent test for indirect clusters
these tests don't make sense without keeping a tunnel open to the
consul port on the first instance for the duration of the entire test.
currently the timeout of this tunnel is on 10 minutes, which may or may
not be long enough if the amount of instances to be started is more than
the amount of workers in the threadpool. when the assimilation finally
happens after all instances have been started a new connection will be
set up to an instance which might not be the first one, which will cause
the shared secret to change between instances and the cluster will fail
to establish consensus.