do not install puppet modules
configuration management should not be part of raptiformica anymore but
belongs in the space of the modules that are installed as raptiformica
modules, so for example vdloo/simulacra will start a jobrunner that will
run the configuration management or vdloo/raptiformica-map which runs
configuration management in a screen detached from the main assimilation
process. this will decouple the overlay network from configuration
management and installing any requirements. breaks vdloo/puppetfiles for
the default vagrant and docker developent instances.