deploy can install modules
note that this does not work if the machine running the deploy is also
the target and is running as root. that will wipe any installed modules
before running the the assimilation. depending on in what order the
machines are ordered in the inventory this could cause some unexpected
behavior.